refactor(fe): refactor form to server side rendering
