+++ frontend/src/app/auth/login/page.tsx
@@ -1,114 +1,24 @@
-"use client";
-
-import { useState } from "react";
+import LoginForm from "@/components/auth/LoginForm";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { useAuthStore } from "@/store/authStore";
-import { authApi } from "@/lib/authApi";
 
 export default function LoginPage() {
-  const router = useRouter();
-  const setAuth = useAuthStore((s) => s.setAuth);
-
-  const [form, setForm] = useState({ email: "", password: "" });
-  const [error, setError] = useState("");
-  const [loading, setLoading] = useState(false);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setForm((prev) => ({ ...prev, [e.target.name]: e.target.value }));
-    setError("");
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-    setError("");
-
-    try {
-      const res = await authApi.login(form);
-      setAuth(res.accessToken, {
-        userId: res.userId,
-        email: res.email,
-        fullName: res.fullName,
-        role: res.role,
-      });
-      router.push("/dashboard");
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Đăng nhập thất bại");
-    } finally {
-      setLoading(false);
-    }
-  };
-
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 px-4">
+    <div className="min-h-screen flex items-center justify-center  px-4">
       <div className="w-full max-w-md">
         <div className="mb-8 text-center">
-          <h1 className="text-2xl font-semibold text-gray-900">Đăng nhập</h1>
-          <p className="mt-1 text-sm text-gray-500">
-            Chưa có tài khoản?{" "}
+          <h1 className="text-2xl font-semibold">Sign In</h1>
+          <p className="mt-1 text-sm text-muted-foreground">
+            Doesn&apos;t have an account?{" "}
             <Link
               href="/auth/register"
-              className="text-indigo-600 hover:underline"
+              className="text-primary hover:underline"
             >
-              Đăng ký
+              Sign Up
             </Link>
           </p>
         </div>
 
-        <div className="bg-white border border-gray-200 rounded-xl p-8">
-          <form onSubmit={handleSubmit} className="space-y-5">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">
-                Email
-              </label>
-              <input
-                type="email"
-                name="email"
-                value={form.email}
-                onChange={handleChange}
-                placeholder="you@example.com"
-                required
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg text-sm
-                           focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent
-                           placeholder:text-gray-400"
-              />
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">
-                Mật khẩu
-              </label>
-              <input
-                type="password"
-                name="password"
-                value={form.password}
-                onChange={handleChange}
-                placeholder="••••••••"
-                required
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg text-sm
-                           focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent
-                           placeholder:text-gray-400"
-              />
-            </div>
-
-            {error && (
-              <p className="text-sm text-red-500 bg-red-50 border border-red-200 rounded-lg px-3 py-2">
-                {error}
-              </p>
-            )}
-
-            <button
-              type="submit"
-              disabled={loading}
-              className="w-full py-2.5 px-4 bg-indigo-600 hover:bg-indigo-700
-                         text-white text-sm font-medium rounded-lg
-                         transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {loading ? "Đang đăng nhập..." : "Đăng nhập"}
-            </button>
-          </form>
-        </div>
+        <LoginForm />
       </div>
     </div>
   );
