+++ frontend/src/app/auth/register/page.tsx
@@ -1,153 +1,21 @@
-"use client";
-
-import { useState } from "react";
+import RegisterForm from "@/components/auth/RegisterForm";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { useAuthStore } from "@/store/authStore";
-import { authApi } from "@/lib/authApi";
 
 export default function RegisterPage() {
-  const router = useRouter();
-  const setAuth = useAuthStore((s) => s.setAuth);
-
-  const [form, setForm] = useState({ email: "", password: "", fullName: "" });
-  const [errors, setErrors] = useState<Record<string, string>>({});
-  const [loading, setLoading] = useState(false);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setForm((prev) => ({ ...prev, [e.target.name]: e.target.value }));
-    setErrors((prev) => ({ ...prev, [e.target.name]: "", general: "" }));
-  };
-
-  const validate = () => {
-    const next: Record<string, string> = {};
-    if (!form.fullName.trim()) next.fullName = "Vui lòng nhập họ tên";
-    if (!form.email.includes("@")) next.email = "Email không hợp lệ";
-    if (form.password.length < 6) next.password = "Mật khẩu ít nhất 6 ký tự";
-    return next;
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    const validation = validate();
-    if (Object.keys(validation).length > 0) {
-      setErrors(validation);
-      return;
-    }
-
-    setLoading(true);
-    try {
-      const res = await authApi.register(form);
-      setAuth(res.accessToken, {
-        userId: res.userId,
-        email: res.email,
-        fullName: res.fullName,
-        role: res.role,
-      });
-      router.push("/dashboard");
-    } catch (err) {
-      setErrors({
-        general: err instanceof Error ? err.message : "Đăng ký thất bại",
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
-
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 px-4">
+    <div className="min-h-screen flex items-center justify-center  px-4">
       <div className="w-full max-w-md">
         <div className="mb-8 text-center">
-          <h1 className="text-2xl font-semibold text-gray-900">
-            Tạo tài khoản
-          </h1>
-          <p className="mt-1 text-sm text-gray-500">
-            Đã có tài khoản?{" "}
-            <Link
-              href="/auth/login"
-              className="text-indigo-600 hover:underline"
-            >
-              Đăng nhập
+          <h1 className="text-2xl font-semibold ">Sign Up</h1>
+          <p className="mt-1 text-sm text-muted-foreground">
+            Already have an account?{" "}
+            <Link href="/auth/login" className="text-primary hover:underline">
+              Log In
             </Link>
           </p>
         </div>
 
-        <div className="bg-white border border-gray-200 rounded-xl p-8">
-          <form onSubmit={handleSubmit} className="space-y-5">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">
-                Họ và tên
-              </label>
-              <input
-                type="text"
-                name="fullName"
-                value={form.fullName}
-                onChange={handleChange}
-                placeholder="Nguyễn Văn A"
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg text-sm
-                           focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent
-                           placeholder:text-gray-400"
-              />
-              {errors.fullName && (
-                <p className="mt-1 text-xs text-red-500">{errors.fullName}</p>
-              )}
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">
-                Email
-              </label>
-              <input
-                type="email"
-                name="email"
-                value={form.email}
-                onChange={handleChange}
-                placeholder="you@example.com"
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg text-sm
-                           focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent
-                           placeholder:text-gray-400"
-              />
-              {errors.email && (
-                <p className="mt-1 text-xs text-red-500">{errors.email}</p>
-              )}
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">
-                Mật khẩu
-              </label>
-              <input
-                type="password"
-                name="password"
-                value={form.password}
-                onChange={handleChange}
-                placeholder="Ít nhất 6 ký tự"
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg text-sm
-                           focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-transparent
-                           placeholder:text-gray-400"
-              />
-              {errors.password && (
-                <p className="mt-1 text-xs text-red-500">{errors.password}</p>
-              )}
-            </div>
-
-            {errors.general && (
-              <p className="text-sm text-red-500 bg-red-50 border border-red-200 rounded-lg px-3 py-2">
-                {errors.general}
-              </p>
-            )}
-
-            <button
-              type="submit"
-              disabled={loading}
-              className="w-full py-2.5 px-4 bg-indigo-600 hover:bg-indigo-700
-                         text-white text-sm font-medium rounded-lg
-                         transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {loading ? "Đang tạo tài khoản..." : "Tạo tài khoản"}
-            </button>
-          </form>
-        </div>
+        <RegisterForm />
       </div>
     </div>
   );
